refactord code for dynamicreconciler
